pyproject.toml
[project]
name = "wiki-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.2",
    "geoalchemy2>=0.18.4",
    "google-genai>=1.69.0",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
    "redis>=5.0.0",
]
